+++ dbdaora/hash/_tests/datastore/conftest.py
@@ -75,7 +75,7 @@ def serialized_fake_entity2():
 
 
 class FakeDatastoreHashRepository(DatastoreHashRepository):
-    entity_name = 'fake'
+    name = 'fake'
     key_attrs = ('other_id', 'id')
     many_key_attrs = ('id',)
     entity_cls = FakeDatastoreEntity