+++ dbdaora/hash/entity.py
@@ -1,7 +1,0 @@
-from typing import Protocol
-
-
-class HashEntity(Protocol):
-    @property
-    def id(self) -> str:
-        ...  # pragma: no cover