+++ dbdaora/sorted_set/entity.py
@@ -1,5 +1,4 @@
-from dataclasses import dataclass
-from typing import Any, Sequence, TypedDict, Union
+from typing import Any, Dict, Protocol, Sequence, TypedDict, Union
 
 from dbdaora.data_sources.memory import RangeOutput
 
@@ -9,12 +8,13 @@ SortedSetInput = Sequence[Union[str, float]]
 SortedSetData = Union[RangeOutput, SortedSetInput]
 
 
-@dataclass(init=False)
-class SortedSetEntity:
-    values: SortedSetData
+class SortedSetEntity(Protocol):
+    @property
+    def values(self) -> SortedSetData:
+        raise NotImplementedError()  # pragma: no cover
 
     def __init__(self, values: SortedSetData, **kwargs: Any):
-        self.values = values
+        raise NotImplementedError()  # pragma: no cover
 
 
 class SortedSetDictEntity(TypedDict):
@@ -23,4 +23,4 @@ class SortedSetDictEntity(TypedDict):
 
 # see https://github.com/python/mypy/issues/4300
 # Entity = Union[SortedSetEntity, SortedSetDictEntity]
-Entity = Union[SortedSetEntity]
+Entity = Union[SortedSetEntity, Dict[str, Any]]