+++ dbdaora/sorted_set/repositories/_tests/test_integration_repository_sorted_set_datastore.py
@@ -1,15 +1,19 @@
+import time
+from typing import TypedDict
+
 import pytest
 from google.cloud import datastore
 from jsondaora import jsondaora
 
-from dbdaora import DatastoreSortedSetRepository, SortedSetDictEntity
+from dbdaora import DatastoreSortedSetRepository, SortedSetData
 from dbdaora.data_sources.fallback.datastore import DatastoreDataSource
 from dbdaora.data_sources.memory.dict import DictMemoryDataSource
 
 
 @jsondaora
-class FakeEntity(SortedSetDictEntity):
+class FakeEntity(TypedDict):
     id: str
+    values: SortedSetData
 
 
 class FakeRepository(DatastoreSortedSetRepository, entity_cls=FakeEntity):
@@ -34,6 +38,7 @@ async def test_should_exclude_all_attributes_from_indexes(repository):
     entity = datastore.Entity(key=key)
     entity.update({'values': values})
     client.put(entity)
+    time.sleep(0.2)
     query = client.query(kind='fake')
     query.add_filter('values', '=', 'v1')
 
@@ -43,6 +48,7 @@ async def test_should_exclude_all_attributes_from_indexes(repository):
     values = [('v1', 1), ('v2', 2)]
     await repository.add(FakeEntity(id='fake', values=values))
 
+    time.sleep(0.2)
     query = client.query(kind='fake')
     query.add_filter('values', '=', 'v1')
     entities = query.fetch()