+++ dbdaora/sorted_set/repositories/datastore.py
@@ -2,7 +2,6 @@ from typing import Any
 
 from google.cloud.datastore import Key
 
-from ..entity import Entity
 from . import SortedSetRepository
 
 
@@ -10,7 +9,7 @@ class DatastoreSortedSetRepository(SortedSetRepository[Key]):
     __skip_cls_validation__ = ('DatastoreSortedSetRepository',)
 
     async def add_fallback(
-        self, entity: Entity, *entities: Entity, **kwargs: Any
+        self, entity: Any, *entities: Any, **kwargs: Any
     ) -> None:
         await super().add_fallback(
             entity, *entities, exclude_from_indexes=('values',)